chore: move lib to shared package
