+++ README.md
@@ -1,6 +1,12 @@
 # powercord-backend
 The backend for [powercord-org/powercord](https://github.com/powercord-org/powercord).
 
+## Repo structure
+ - `packages/api`: REST API
+ - `packages/web`: Web UI & HTTP server for pre-rendering
+ - `packages/boat`: Discord bot for the Powercord server
+ - `packages/crapcord`: Library for interfacing with the Discord API
+
 ## License
 This software is licensed under MIT license, with a few exceptions:
  - `packages/web/src/components/docs/Markdown.tsx` is courtesy of Borkenware as part of Spoonfeed, licensed under BSD-3-Clause
