+++ packages/boat/package.json
@@ -7,7 +7,7 @@
   "private": true,
   "type": "module",
   "engines": {
-    "node": ">= 14",
+    "node": ">= 15",
     "pnpm": ">= 6"
   },
   "scripts": {
