+++ packages/crapcord/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@powercord/api",
-  "version": "2.0.0",
-  "description": "Powercord backend API",
+  "name": "@powercord/crapcord",
+  "version": "0.1.0",
+  "description": "Minimal Discord API wrapper",
   "repository": "git@github.com:powercord-org/powercord-backend.git",
   "license": "MIT",
   "private": true,
@@ -17,22 +17,15 @@
     "lint": "eslint --ext ts --ext tsx src"
   },
   "dependencies": {
-    "@borkenware/spoonfeed": "workspace:*",
-    "@powercord/types": "workspace:*",
-    "fastify": "^3.20.2",
-    "fastify-auth": "^1.1.0",
-    "fastify-cookie": "^5.3.1",
-    "fastify-mongodb": "^4.1.0",
-    "fastify-raw-body": "^3.1.0",
-    "fastify-tokenize": "^1.3.1",
-    "mongodb": "^4.1.0",
-    "node-fetch": "^2.6.1"
+    "tiny-typed-emitter": "^2.1.0",
+    "ws": "^8.2.0"
   },
   "devDependencies": {
     "@types/node": "^16.6.1",
-    "@types/node-fetch": "^2.5.12",
+    "@types/ws": "^7.4.7",
     "@typescript-eslint/eslint-plugin": "^4.29.2",
     "@typescript-eslint/parser": "^4.29.2",
+    "discord-api-types": "^0.22.0",
     "eslint": "^7.32.0",
     "eslint-plugin-deprecation": "^1.2.1",
     "tsc-watch": "^4.4.0",
