+++ packages/crapcord/tsconfig.json
@@ -1,7 +1,8 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "outDir": "./dist"
+    "outDir": "./dist",
+    "typeRoots": [ "./node_modules/@types" ]
   },
   "include": [ "./src/**/*" ]
 }
