+++ packages/web/package.json
@@ -6,7 +6,7 @@
   "license": "MIT",
   "private": true,
   "engines": {
-    "node": ">= 14",
+    "node": ">= 15",
     "pnpm": ">= 5.13"
   },
   "scripts": {
